rename 'panel' to 'launcher.
