+++ docs/reference/api/tac.gui.launcher.api.resources.rst
@@ -1,0 +1,30 @@
+tac.gui.launcher.api.resources package
+======================================
+
+Submodules
+----------
+
+tac.gui.launcher.api.resources.agents module
+--------------------------------------------
+
+.. automodule:: tac.gui.launcher.api.resources.agents
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+tac.gui.launcher.api.resources.sandboxes module
+-----------------------------------------------
+
+.. automodule:: tac.gui.launcher.api.resources.sandboxes
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: tac.gui.launcher.api.resources
+   :members:
+   :undoc-members:
+   :show-inheritance:
