+++ docs/reference/api/tac.gui.launcher.api.rst
@@ -1,0 +1,17 @@
+tac.gui.launcher.api package
+============================
+
+Subpackages
+-----------
+
+.. toctree::
+
+   tac.gui.launcher.api.resources
+
+Module contents
+---------------
+
+.. automodule:: tac.gui.launcher.api
+   :members:
+   :undoc-members:
+   :show-inheritance:
