+++ docs/reference/api/tac.gui.launcher.forms.rst
@@ -1,0 +1,30 @@
+tac.gui.launcher.forms package
+==============================
+
+Submodules
+----------
+
+tac.gui.launcher.forms.agent module
+-----------------------------------
+
+.. automodule:: tac.gui.launcher.forms.agent
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+tac.gui.launcher.forms.sandbox module
+-------------------------------------
+
+.. automodule:: tac.gui.launcher.forms.sandbox
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: tac.gui.launcher.forms
+   :members:
+   :undoc-members:
+   :show-inheritance:
