+++ docs/reference/api/tac.gui.launcher.rst
@@ -1,0 +1,46 @@
+tac.gui.launcher package
+========================
+
+Subpackages
+-----------
+
+.. toctree::
+
+   tac.gui.launcher.api
+   tac.gui.launcher.forms
+
+Submodules
+----------
+
+tac.gui.launcher.app module
+---------------------------
+
+.. automodule:: tac.gui.launcher.app
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+tac.gui.launcher.config module
+------------------------------
+
+.. automodule:: tac.gui.launcher.config
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+tac.gui.launcher.home module
+----------------------------
+
+.. automodule:: tac.gui.launcher.home
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: tac.gui.launcher
+   :members:
+   :undoc-members:
+   :show-inheritance:
