--- docs/reference/api/tac.gui.panel.api.resources.rst
@@ -1,30 +1,0 @@
-tac.gui.panel.api.resources package
-===================================
-
-Submodules
-----------
-
-tac.gui.panel.api.resources.agents module
------------------------------------------
-
-.. automodule:: tac.gui.panel.api.resources.agents
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-tac.gui.panel.api.resources.sandboxes module
---------------------------------------------
-
-.. automodule:: tac.gui.panel.api.resources.sandboxes
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: tac.gui.panel.api.resources
-   :members:
-   :undoc-members:
-   :show-inheritance:
