--- docs/reference/api/tac.gui.panel.api.rst
@@ -1,17 +1,0 @@
-tac.gui.panel.api package
-=========================
-
-Subpackages
------------
-
-.. toctree::
-
-   tac.gui.panel.api.resources
-
-Module contents
----------------
-
-.. automodule:: tac.gui.panel.api
-   :members:
-   :undoc-members:
-   :show-inheritance:
