--- docs/reference/api/tac.gui.panel.forms.rst
@@ -1,30 +1,0 @@
-tac.gui.panel.forms package
-===========================
-
-Submodules
-----------
-
-tac.gui.panel.forms.agent module
---------------------------------
-
-.. automodule:: tac.gui.panel.forms.agent
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-tac.gui.panel.forms.sandbox module
-----------------------------------
-
-.. automodule:: tac.gui.panel.forms.sandbox
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: tac.gui.panel.forms
-   :members:
-   :undoc-members:
-   :show-inheritance:
