--- docs/reference/api/tac.gui.panel.rst
@@ -1,46 +1,0 @@
-tac.gui.panel package
-=====================
-
-Subpackages
------------
-
-.. toctree::
-
-   tac.gui.panel.api
-   tac.gui.panel.forms
-
-Submodules
-----------
-
-tac.gui.panel.app module
-------------------------
-
-.. automodule:: tac.gui.panel.app
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-tac.gui.panel.config module
----------------------------
-
-.. automodule:: tac.gui.panel.config
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-tac.gui.panel.home module
--------------------------
-
-.. automodule:: tac.gui.panel.home
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: tac.gui.panel
-   :members:
-   :undoc-members:
-   :show-inheritance:
