+++ docs/reference/api/tac.gui.rst
@@ -7,7 +7,7 @@ Subpackages
 .. toctree::
 
    tac.gui.dashboards
-   tac.gui.panel
+   tac.gui.launcher
 
 Submodules
 ----------
