+++ tac/gui/launcher/__init__.py
@@ -25,23 +25,6 @@ In particular, it provides REST methods to start/stop a sandbox and an agent, al
  user to easily change the parameters.
 """
 
-# ------------------------------------------------------------------------------
-#
-#   Copyright 2018-2019 Fetch.AI Limited
-#
-#   Licensed under the Apache License, Version 2.0 (the "License");
-#   you may not use this file except in compliance with the License.
-#   You may obtain a copy of the License at
-#
-#       http://www.apache.org/licenses/LICENSE-2.0
-#
-#   Unless required by applicable law or agreed to in writing, software
-#   distributed under the License is distributed on an "AS IS" BASIS,
-#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#   See the License for the specific language governing permissions and
-#   limitations under the License.
-#
-# ------------------------------------------------------------------------------
 import logging
 import os
 from queue import Empty
@@ -49,8 +32,8 @@ from threading import Thread
 
 from flask import Flask
 
-from tac.gui.panel import home, api
-from tac.gui.panel.api.resources.sandboxes import sandbox_queue, SandboxRunner
+from tac.gui.launcher import home, api
+from tac.gui.launcher.api.resources.sandboxes import sandbox_queue, SandboxRunner
 
 logger = logging.getLogger(__name__)
