+++ tac/gui/launcher/api/__init__.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+
 # ------------------------------------------------------------------------------
 #
 #   Copyright 2018-2019 Fetch.AI Limited
@@ -17,4 +18,17 @@
 #
 # ------------------------------------------------------------------------------
 
-"""Define all the REST resources for the panel APIs."""
+"""Define the REST APIs for the launcher app."""
+from flask_restful import Api
+
+from .resources.sandboxes import SandboxList, Sandbox
+from .resources.agents import Agent
+
+
+def create_api(app):
+    """Wrap the Flask app with the Flask-RESTful Api object."""
+    api = Api(app, prefix='/api')
+
+    api.add_resource(SandboxList, "/sandboxes")
+    api.add_resource(Sandbox, "/sandboxes/<int:sandbox_id>")
+    api.add_resource(Agent, "/agent")
