+++ tac/gui/launcher/app.py
@@ -18,4 +18,10 @@
 #
 # ------------------------------------------------------------------------------
 
-"""Implement forms to display in the panel GUI."""
+"""The main entrypoint for the launcher app."""
+
+from tac.gui.launcher import create_app
+
+if __name__ == '__main__':
+    app = create_app()
+    app.run("127.0.0.1", 5000, debug=True, use_reloader=False)
