+++ tac/gui/launcher/forms/__init__.py
@@ -18,10 +18,4 @@
 #
 # ------------------------------------------------------------------------------
 
-"""The main entrypoint for the panel app."""
-
-from tac.gui.panel import create_app
-
-if __name__ == '__main__':
-    app = create_app()
-    app.run("127.0.0.1", 5000, debug=True, use_reloader=False)
+"""Implement forms to display in the launcher GUI."""
