+++ tac/gui/launcher/static/js/launcher.js
