+++ tac/gui/launcher/templates/includes/navbar.html
@@ -6,7 +6,7 @@
   <div class="collapse navbar-collapse" id="navbarNav">
     <ul class="navbar-nav">
       <li class="nav-item active">
-        <a class="nav-link" href="/panel">Panel<span class="sr-only">(current)</span></a>
+        <a class="nav-link" href="/launcher">Launcher<span class="sr-only">(current)</span></a>
       </li>
       <li class="nav-item">
         <a class="nav-link" href="/grid-search">Grid Search</a>
