+++ tac/gui/launcher/templates/launcher.html
@@ -45,5 +45,5 @@
 </div>
 {% endblock %}
 {% block scripts %}
-<script src="/static/js/panel.js"></script>
+<script src="/static/js/launcher.js"></script>
 {% endblock %}
